apps/backend/pyproject.toml
[project]
name = "doc-extractor"
version = "0.1.0"
description = "Document data extraction using Mistral AI"
dependencies = [
    "mistralai>=1.2.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0", # Required for PDF processing
    "pydantic>=2.0.0",
    "pymupdf>=1.23.0",
    "pydantic-settings>=2.10.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[dependency-groups]
dev = [
    "gitingest>=0.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*"]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-cov",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build]
only-packages = true
